requirements.txt
certifi==2019.3.9
Django==2.2.2
django-debug-toolbar==2.0
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
